requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
Jinja2==3.1.2
SQLAlchemy==2.0.23
python-dotenv==1.0.0
email-validator==2.1.0
firebase-admin==6.2.0
google-cloud-firestore==2.13.1
google-auth==2.23.4
gunicorn==21.2.0
requests>=2.30.0
